DPG: allow hyperloop running on derived data
- Separate split and non split tasks
- Add more histos for MC
- Fix process function of MC
